philo

+++ common-core/3-circle/philosophers/srcs/utils/sleep_utils.c
@@ -6,7 +6,7 @@
 /*   By: mivogel <mivogel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/02 14:14:55 by mivogel           #+#    #+#             */
-/*   Updated: 2025/04/08 14:44:24 by mivogel          ###   ########.fr       */
+/*   Updated: 2025/04/09 13:54:29 by mivogel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ int	ft_get_time(void)
 	return ((time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
 
-void	ft_usleep(size_t time)
+void	ft_usleep(int time)
 {
-	size_t start;
+	int	start;
 
 	start = ft_get_time();
 	while (ft_get_time() - start < time)
